Make QQmlExtensionPlugin documentation clearer.
I spent too long wondering why my plugin wasn't being loaded, until I
realized that the directory I was installing the qmldir into was not
named the same as the module. This is already documented in the
"Module Definition qmldir Files" documentation:

http://doc.qt.io/qt-5/qtqml-modules-qmldir.html#contents-of-a-module-definition-qmldir-file

But as I was writing a plugin, I was following this documentation:

http://doc.qt.io/qt-5/qtqml-modules-cppplugins.html

Where it was not at all obvious that they should be named the same.

Change-Id: I3e20bc31f8b42c7141b4c22c8cb1750ba9782971
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>