GifRecorder: add a high quality option.
byzanz-record will heavily optimize colors in certain GIFs. The GIF
generated for BusyIndicator, for example, lost all of its
"transparency" that it has in the source image.

The man page says that the .flv format is lossless, so when
setHighQuality(true) is called, we record a video and then
convert it to a bunch of frames with avconv, then to a GIF with
convert.

Finally, to further optimize the GIF, we could run it through gifsicle,
so I haven't updated any GIFs in the docs yet.

Change-Id: I0ccc5e2258d7c6dac52e464bcd6fb61128541e82
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>