Add documentation to module page, and remove overview page.
The overview page doesn't seem necessary; it's one more click that can be
avoided by presenting the information that it contains on the index page.

I didn't categorise the QML types as was done for Qt Quick Controls,
because we don't have enough types to fill most of those categories.

Change-Id: I8456fe219eb69983229712b0d9f5eef1f3a83ada
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>