GifRecorder: don't assume QQuickView
This allows us to use windows as root items, which means we can have
complete examples - not just a snippet without imports. With this, we
can have runnable snippets under each GIF, without the need for the
user to add code to get it to run.

This change is also necessary to generate GIFs for Panel-based items
like Menu, which require an ApplicationWindow.

Change-Id: I66fec30cde8d719be35766f8c917cd2a06b09127
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>