QTextEdit/QQuickTextEdit undo issue - Part #1
This change to the qquicktextedit testcase will match updated
functionality for undo (made in qtbase) in the following scenario:

If a user selected the text "foo" and typed "bar", upon pressing undo,
the text would change to "b". This is incorrect and does not match the
functionality of QTextEdit or the default behaviours of
Windows/OSX/Ubuntu. When pressing undo on QTextEdit, the text would
change back to "foo".

The changes to the testcase is basically removing one step in the undo
test data that does not happen anymore. Since the change to undo was
made in qtbase, I need to disable the undo test cases using the
BLACKLIST until the qtbase change is in. Then there will be another
review to re-enable it.

Task-number: QTBUG-38825
Change-Id: I60e799d546945aac89634b49b2e1dc60a2a68935
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>