Fix StackView: pushed items still showed up after being popped
This is a partial revert of 6ee9612. The idea was to keep items visible
until both transitions had finished.

   StackView {
       pushExit: null
       pushEnter: Transition { ... }
   }

When both transitions had finished, all stack elements (except the
current one) were hid. The problem was that in pop & replace operations
the exiting item was no longer part of the stack (because it had been
already popped out) and therefore was never hid.

This change restores the original (intended) behavior that a null
transition would equal to an immediate operation. This keeps the code
simpler, and makes the behavior consistent in a sense that all items
are hidden as soon the respective transition finishes. PauseAnimation
can be used to wait if desired.

Change-Id: Ia9e78506754bdc18949676f061dbfff28a7128b7
Task-number: QTBUG-47317
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>