Fix compilation with ICC on Windows
QV4::Managed is not copyable and its default constructor is deleted.
However, it and classes derived from it are exported, which on Windows
means the compiler will instantiate all possible functions and add to
the DLL. ICC on Windows, unlike MSVC, attempts to instantiate the
default constructor of the derived classes (like CallContext) and then
the build fails due to the deleted Managed() constructor.

Instead, use V4_MANAGED to mark each and every managed class as non-
default-constructible and non-copyable. Only one note: the V4_MANAGED
macro in QV4::Managed itself takes different parameters, so it needs to
be slightly different.

Task-number: QTBUG-48063
Change-Id: I42e7ef1a481840699a8dffff140007c65a7a35db
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>