QQuickWidget: do not render when there is no fbo due to no size
It is not invalid to have a widget with a 0,0 size. The FBO logic
already handles this: no FBO is created in this case, it will be done
on a subsequent resize. However, we can still get to render() which
is not currently prepared to handle this case. Correct this.

Task-number: QTBUG-47588
Task-number: QTBUG-48059
Change-Id: Idd674c536847e5f293fec5a40aa3038600473c27
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>