Avoid crashing when not having a screen for the context
Disconnecting the screen to which the context belongs results in a null
QOpenGLContext::screen(). Fall back to qguiapp's devicePixelRatio() in
this case (giving the highest dpi in the system) due to the lack of any
other option.

This will fix the crashes when rendering offscreen (QQuickWidget) and
using native text rendering. Without this the render attempt after
disconnecting the screen with which the context was created would
crash due to accessing the already destroyed screen instance.

Task-number: QTBUG-42803
Change-Id: Ie5889f59e5d66d09329b3d5fb8e1365d14d5b8fc
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>