Add ExclusiveGroup::checkables
Exclusive checkables typically share the same parent item. For example,
a row or column positioner or layout. This change makes it convenient
to handle such cases so that one can add the whole list of children to
an exclusive group, without the need of creating any attached objects.

    Column {
        id: column

        ExclusiveGroup {
            id: eg
            checkables: column.children
        }

        Switch { id: s1; text: qsTr("First") }
        Switch { id: s2; text: qsTr("Second") }
        Switch { id: s3; text: qsTr("Third") }
    }

Last but not least, this allows breaking the C++ dependency between
QQuickTabBar and QQuickExclusiveGroup.

Change-Id: Ic75065b87ef8fa532f77118d740f04484994bbf0
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>