Ensure press-and-hold event keeps selection
This concerns TextField and TextArea.

It is an almost universal UX pattern on touch platforms where
the user long presses to pop the context menu up. In many
cases the context menu is used for copy and cut operations,
which means that poping it up should keep the selection.

The implementation works by not forwarding the initial mouse
press event to the parent class until we're sure it's not going
to be a long press. If the long press timer is cancelled for any
reason, we will then send the delayed mouse press event to the
parent class followed by whichever event triggered the cancellation.

Auto-tests refactored and updated.

Change-Id: If3aa8075f07a80929f4bd723895d9599bf8d169e
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>