Add support for mapping between item's and global screen coordinates
Added to Item mapToGlobal() and mapFromGlobal() methods, that maps from
item's coordinate system to global screen coordinates and vise versa
and accessible from QML.

This allow to align window (which is a global object) in relation to
Qt Quick item object. For example, this may be helpful to add a popup to
a Qt Quick component.

[ChangeLog][QtQuick][QQuickItem] Added support for mapping item's
coordinates to global screen coordinates and vise versa.

Task-number: QTBUG-28668
Change-Id: I0a90838f26c53b16d7a76953577492a942674572
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>