Doc: replaced \target with \keyword if at top of page
A \target whose purpose is to link to the top of a
page (and not to a section within a page) works better
as a \keyword, because \target generates a
new html anchor which, in this case, is not tied to
any title element on the page.

A \keyword links to the page itself, as expected.

Change-Id: Ic7c83efc5066674aba3d65164bf2cf53697df4d6
Task-number: QTBUG-48482
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>
Reviewed-by: Martin Smith <martin.smith@digia.com>