Fixed TabBar reset currentIndex to -1 on completion
TabBar resets currentIndex to -1 on completion, even if it has been
explicitly set to a proper value. This is due to a calling
updateCurrent() on completion.

As updateCurrent() handle cases, where tab clicked by user, and proper
index handled by QQuickContainer, there is no need to call this function
on TabBar completion.

Task-number: QTBUG-48624
Change-Id: Ied316af2bfb8983ea1d5947544cc8399aaf7eb84
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>