Give SquareButton types a side property to mediate being square.
We shouldn't be setting width and height for each square; we should
only need to set their side, which determines their width and height -
so that they're actually square !

More pertinently, this could later serve as an illustration of how to
use properties and bindings.  Ideally, we'd do this in some way that
actually ensures setting width or height changes side; but the simple
approach will do for now.

Change-Id: I0df89e11221dde5d1d2930c4a779b59cf0d46654
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>