Remove line numbers for statements without side-effects.
When debug mode is not enabled, line numbers are only used in order to
generate usable stack traces. Therefore statements that cannot
fail/throw do not need line numbers associated with them. Of course,
when debug mode is enabled, this is not applicable.

The effect is that the generated code for such statements shrinks by 2
loads and 1 store.

Change-Id: I4ec425dd20b56043e8ca0e4c68afc683eb9b50e7
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>