QQuickItem: correct focus reason for key navigation
Before we used Qt::OtherFocusReason for normal key navigation, such
as up/down/left/right. But in style world, the reason is critical
for focus change.

Now we follow the way in widgets world, use Qt::TabFocusReason and
Qt::BacktabFocusReason for down/up and right/left(or reversed if
mirrored).

Task-number: QTBUG-41801
Change-Id: I99e97ea4ced2cef9b3280c578c3a463de8fe1727
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>