Make it possible to change sources of ShaderEffect again.
92433623b31388e2e8c4d532033dad6189eaab24 introduced a bug where
an unused material would clear its typeid pointer. This resulted
in that a changed shadereffect would in all likelyhood get the
same pointer for its changed shader which would result in using
the same GL shader inside the renderer.

This change rewrites the logic so that the cache is per render
thread and the cache is cleaned up along with other GL/SG resources
as part of scene graph invalidation.

Task-number: QTBUG-48856
Change-Id: Id2feb14f584b5f5c176e8176cc9f1b54abd0d079
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>