Move the internal ExclusiveGroup from TabBar to Container
TabBar needs to support PathView as a content item (for the Universal
style). PathView creates a highlight item, which wrongly ends up being
added as a content child. It shows up as an extra empty TabBar item, and
leads to a warning from ExclusiveGroup::addCheckable().

An exclusive group instance is lazily created only for those containers
that need it. This allows us to simplify the virtual hooks for added and
removed items, whereas otherwise we would had had to add more complex
hooks (before _and_ after adding an item) to be able to control/ignore
the highlight somehow from the TabBar subclass.

Change-Id: I2c35a1dd80541bc6b30e713a760277940a5486ed
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>