TabBar: fix binding loop on currentIndex
TabBar is a container of exclusive items aka TabButtons. In this case
we let ExclusiveGroup manage the currently checked item, and just sync
the current index when appropriate. In the case of a SwipeView, which
contains arbitrary non-checkable items, we track the current index of
the content item. However, we should avoid doing both at the same time,
because it leads to a race between content item's current index and
exclusive group's current item.

Change-Id: Ib2d4dfe0d22d883be49b03a3b2311bdea7d937d9
Task-number: QTBUG-48993
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>