Remove PageIndicator::color
The color property was added in e828b9514. The idea was that it would
be convenient to match the indicator against a specific background color,
and that the property would be designable. However, the color property
is basically impossible to support for styles that have specific color
palettes for dark and light themes. => Implement a custom delegate if
customization is required.

Change-Id: I044bbe560a59362c585c801271cd2852456287b7
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>