Make tst_tabbar pass with a PathView
The upcoming universal style uses PathView as a content item for
TabBar. Compared to ListView, PathView has some differences in the
hanling of current index. For example, the current index of an empty
PathView is reseted to 0. This patch ensures that Container retains
the correct current index during insert/move/remove operations, no
matter which content item is used.

Change-Id: I385121d75a410e27d7bef9f5042a51948c7d7db3
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>