Refactor press and hold tests
We can't rely on timers in auto tests. They are not fired accurately
under heavy CPU load => random CI failures. This change make the press
and hold tests independent of timeouts.

The idea in a nutshell: the test creates two control instances, and
sends press events to both of them. Both controls run their timers and
emit pressAndHold() - unless canceled. The test waits as long as it
takes for the second control to emit pressAndHold(). At that point we
can reliably assume that the first control would have already emitted
the signal, if it was not canceled as appropriate.

Change-Id: I96aaf4d204db368171209271a4a0fa2ff15cc323
Task-number: QTBUG-47963
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>