Private export QQuickPen
When including QtQuick headers outside of QtQuick, MSVC fails to link
the staticMetaObject of any unexported Q_OBJECT class that it sees.

QQuickPen is the only autotest-exported type in qquickrectangle_p.h.
All other types in the same header (QQuickRectangle, QQuickGradient,
and QQuickGradientStop) are private exported. Because of the MSVC
linking behavior described above, those private exports are worthless
if the last remaining type is not exported.

Change-Id: Ife1bf466fa46a1b86ddf8ff76ab73c1beaa71acc
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>